+++ packages/public-scoped-sdk/index.js
@@ -1,0 +1,10 @@
+exports.PUBLIC_SCOPED_SDK = [
+  "cmdb",
+  "micro-app",
+  "monitor",
+  "msgsender",
+  "next-builder",
+  "object-store",
+  "permission",
+  "user-service",
+];