+++ packages/sdk-scripts/src/constants.ts
@@ -1,1 +1,0 @@
-export const PUBLIC_SCOPED_SDK = ["cmdb", "micro-app", "user-service"];