+++ dll/ace/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.83](https://github.com/easyops-cn/next-core/compare/@next-dll/ace@2.0.82...@next-dll/ace@2.0.83) (2021-01-30)
+
+**Note:** Version bump only for package @next-dll/ace
+
+## [2.0.82](https://github.com/easyops-cn/next-core/compare/@next-dll/ace@2.0.81...@next-dll/ace@2.0.82) (2021-01-28)
+
+**Note:** Version bump only for package @next-dll/ace
+
+## [2.0.81](https://github.com/easyops-cn/next-core/compare/@next-dll/ace@2.0.80...@next-dll/ace@2.0.81) (2021-01-28)
+
+**Note:** Version bump only for package @next-dll/ace
+
 ## [2.0.80](https://github.com/easyops-cn/next-core/compare/@next-dll/ace@2.0.79...@next-dll/ace@2.0.80) (2021-01-26)
 
 **Note:** Version bump only for package @next-dll/ace