+++ dll/editor-bricks-helper/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.16](https://github.com/easyops-cn/next-core/compare/@next-dll/editor-bricks-helper@0.4.15...@next-dll/editor-bricks-helper@0.4.16) (2021-01-30)
+
+**Note:** Version bump only for package @next-dll/editor-bricks-helper
+
+## [0.4.15](https://github.com/easyops-cn/next-core/compare/@next-dll/editor-bricks-helper@0.4.14...@next-dll/editor-bricks-helper@0.4.15) (2021-01-28)
+
+**Note:** Version bump only for package @next-dll/editor-bricks-helper
+
+## [0.4.14](https://github.com/easyops-cn/next-core/compare/@next-dll/editor-bricks-helper@0.4.13...@next-dll/editor-bricks-helper@0.4.14) (2021-01-28)
+
+**Note:** Version bump only for package @next-dll/editor-bricks-helper
+
 ## [0.4.13](https://github.com/easyops-cn/next-core/compare/@next-dll/editor-bricks-helper@0.4.12...@next-dll/editor-bricks-helper@0.4.13) (2021-01-26)
 
 **Note:** Version bump only for package @next-dll/editor-bricks-helper