+++ dll/react-dnd/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.21](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.20...@next-dll/react-dnd@0.1.21) (2021-01-30)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.20](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.19...@next-dll/react-dnd@0.1.20) (2021-01-28)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.19](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.18...@next-dll/react-dnd@0.1.19) (2021-01-28)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
 ## [0.1.18](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.17...@next-dll/react-dnd@0.1.18) (2021-01-26)
 
 **Note:** Version bump only for package @next-dll/react-dnd