+++ packages/brick-container/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.30](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@2.15.29...@next-core/brick-container@2.15.30) (2021-01-30)
+
+**Note:** Version bump only for package @next-core/brick-container
+
+## [2.15.29](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@2.15.28...@next-core/brick-container@2.15.29) (2021-01-28)
+
+**Note:** Version bump only for package @next-core/brick-container
+
+## [2.15.28](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@2.15.27...@next-core/brick-container@2.15.28) (2021-01-28)
+
+**Note:** Version bump only for package @next-core/brick-container
+
 ## [2.15.27](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@2.15.26...@next-core/brick-container@2.15.27) (2021-01-26)
 
 **Note:** Version bump only for package @next-core/brick-container