+++ packages/brick-kit/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.32.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.31.2...@next-core/brick-kit@2.32.0) (2021-01-30)
+
+### Features
+
+- add permissionsPreCheck to route and brick ([4df60fd](https://github.com/easyops-cn/next-core/commit/4df60fd))
+
+## [2.31.2](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.31.1...@next-core/brick-kit@2.31.2) (2021-01-28)
+
+**Note:** Version bump only for package @next-core/brick-kit
+
 ## [2.31.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.31.0...@next-core/brick-kit@2.31.1) (2021-01-26)
 
 ### Bug Fixes